Rendering Done?
I though that the mesh was 'Blipping' because it was a little too long. Turns out it was because it was done in the Presentation Step, which meant I created a job that completed after the sync point but wanted to finish before the sync point. Moved it to initialization, should solve the issue for now.

Among other things, made some terrible decisions about storing/referencing data from Jobs. Its a clustertruck to work with, prototype with, and I already regret it. A write up is most likely going to happen about everything I hate how I implimented.

But, the next step I'm going to undertake is a simple editor, because reserving qweasd for movement (and right click) interfered with my original m1/m2 system of clickping to alter or remove blocks, making that easier to do and understand is a must.