Made Life Worse
There is definately going to be a rewrite in the future for registries.

Looking at how I made my QOL improvements. It just looks terrible and unintuitive.

But, I've got more important things to do RN. Namely moving the logic of updating HiddenFaces to a system. Systems are another issue I should talk about, they work, but it seems like a roundabout way of doing this.