Forgot To Add Culling System
Raycaster no longer manually fixes faces.

So, as I was running into before, processing the entire chunk when one piece changes is expensive (100 ms for 32K pieces of data). Either we blow even more memory to keep a cache of last version and then run a parallel job to find changed, or we have a queue that is updated when something changes... using ECS i only know how to do option one, but since that doubles our data size, I'll be looking for alternatives.

Might end up shrinking chunks to byte size, ore slashing chunkSize in half (16^3 is like 4096, give or take a power of 2 or so.