Cleaned Up An Pretified Notes From Part 0
Kept most of the content the same, since it's meant for me to return to it and talk about it later.

Part 0, by the way, is before I kept track of when I wrote the notes, I know i could look at the git history, but this is easier and really simple.