Fixed System Bug, Learned About Chunk Versioning
SO. Definately should do a writeup about this, after glancing at the profielr I noticed that multiple jobs were being scheduled per frame. I was pretty sure this was impossible since we only request one chunk per frame.

As it turns out (and honeslty it makes sense) is that the version is for the Chunk, not the entity. It makes sense, since EntityVersions would probably only keep track of whether the entity was written to, which would probably be CONSTANTLY. Additionally, under the hood I bet it would be writing to random locatiosn constantly.

So, once I realized this, I realized UGH i need to impliment my own entity changed system. Since the system doesnt want to run on an entity that hasn't changed (because thats alot of data to process).

Imagine if Voxels were entities, this would be a nightmare for memory management. But since Chunks are entities, we just get away with checking if the chunk changed.

My code works on a per component basis instead of one per chunk, this prevents us from having to run all systems, even though (more than likely) it will have to run all systems.

I guess this means that IF all (or most all) systems run, than we can just move to aa single ChunkVersion component instead.