Update NativeValue.cs
Native Value now Mirrors NativeArray (which unfortunately for future me, means we now allow unsafe code).

This means we can use most features Native Types can use.
Incase anyone is wondering, if you want to return a signle type, you have to create a nativearray with a size of one, this removes that 'special return array' with a custom type meant to do the same thing. It's a QOL improvement.