Attempted Chunk Mesh Gen Solution
This tries implimenting a MeshEntity intermediate type to generate meshes, it had some good ideas (and might be useful later)

However, after thinking how ridicoulous a MeshEntity is, it would make more sense to use a NativeHashMap, and a queue to check if Jobs are complete before creating a mesh.

Still, commiting to save the changes i've made. But this cant be run.