Code Cleanup & Some Input Dependency Fixes
Unrelated, due to chunkgen, it looks like we get up to 30fps now. So maybe start moving test system code to the chunkgen system to get that wicked 60 fps. (yeah Mesh Gen is still a problem, but when we can greedy mesh its not)

Related, we have some cyclic job dependencies, get on that.