"It Just Works"
While  I had to rewrite or move alot of code (mostly to prevent problems with handles) i think i like where this is going. By using the 'world position' and the ChunkMap we seperate the hassle of getting the chunk and whatnot. We dont error if the chunk isnt loaded (which kinda sucks) and the entities are not deterministic (basically, HUGE RED FLAG) but there are other factors for nondeterminism so im not worried at the moment. (Multiplayer lay means thinks aren't deterministic clientside anyways.)

But the biggest reason i like this new system, as the man todd howard said, "It Just Works"